requirements.txt
flask
python-dotenv
pyjwt
flask-cors
requests
openai
